fix: '친구 목록 페이지' 데이터 API와 통일 및 변환 처리[S12P11A808-120]

+++ app/src/main/java/com/example/gogoma/data/model/Friend.kt
@@ -9,5 +9,5 @@ data class Friend (
     val friendId : Int,
     val name : String,
     val profileImage : String?,
-    val totalDistance: Double
+    val totalDistance: Int
 ) : Parcelable
